+++ app/controllers/channels_controller.rb
@@ -4,7 +4,7 @@ class ChannelsController < ApplicationController
   # GET /channels
   # GET /channels.json
   def index
-    @channels = Channel.all.page(params[:page])
+    @channels = Channel.all
     render json: @channels
   end
